Firebase + swagger authentication
Created authentication providers and filters for Firebase and Swagger authentication.
Firebase auth requires a header titled "Firebase_Authorization" that contains the Firebase idToken, and allows full access to the endpoints.
Swagger auth is a dev only authentication method that enables Swagger to access the endpoints, provided a username and password ("admin", "proconnect")

The filters are added in the ApiSecurityBuilderConfigurer class